+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 
+ - Node graph mutation (stroke, stroke-width, fill)
  - Added npm build scripts
  - Added userOptions on initialization
  - Added UMD library definition (full d3 at the moment)